LUCENE-2633: fix int/long overflow issue in packed ints impl
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@995409 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/util/packed/Packed32.java
@@ -186,7 +186,7 @@ class Packed32 extends PackedInts.ReaderImpl implements PackedInts.Mutable {
    * @return the value at the given index.
    */
   public long get(final int index) {
-    final long majorBitPos = index * bitsPerValue;
+    final long majorBitPos = (long)index * bitsPerValue;
     final int elementPos = (int)(majorBitPos >>> BLOCK_BITS); // / BLOCK_SIZE
     final int bitPos =     (int)(majorBitPos & MOD_MASK); // % BLOCK_SIZE);
 
@@ -198,7 +198,7 @@ class Packed32 extends PackedInts.ReaderImpl implements PackedInts.Mutable {
 
   public void set(final int index, final long value) {
     final int intValue = (int)value;
-    final long majorBitPos = index * bitsPerValue;
+    final long majorBitPos = (long)index * bitsPerValue;
     final int elementPos = (int)(majorBitPos >>> BLOCK_BITS); // / BLOCK_SIZE
     final int bitPos =     (int)(majorBitPos & MOD_MASK); // % BLOCK_SIZE);
     final int base = bitPos * FAC_BITPOS;
